==> 02_all_states_plots.ipynb <==
# All States
----
This contains the merged DataFrames of California, Texas, New York and Florida.

Data is limited to:
 - Years 2014-2017
 - People over the 20<sup>th</sup> percentile and under the 90<sup>th</sup> percentile by state
 - People in the few major Metro Areas by state

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import json
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.impute import SimpleImputer
from sklearn.preprocessing import OneHotEncoder, MinMaxScaler
from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer


def compress_dataframe(df, dictionary):
    df = df.copy(deep=True)
    
    for col in dictionary.keys():
        df[col] = df[col].map(dictionary[col])
    return df

def reduce_large_strings(df):
    to_replace = ["Information not provided by applicant in mail, Internet, or telephone application",
                  "Native Hawaiian or Other Pacific Islander",
                  "One-to-four family dwelling (other than manufactured housing)"]
    replace_with = ["Information not Provided",
                    "Native Hawaiian/Pacific Islander",
                    "1-4 Family House"]
    for col in df.columns:
        if df[col].dtype == object:
            for i, _ in enumerate(to_replace):
                df[col] = np.where(df[col] == to_replace[i], replace_with[i], df[col])            
    return df

# Code from Murmel on Stack Exchange
# https://stackoverflow.com/questions/1450957/pythons-json-module-converts-int-dictionary-keys-to-strings
def jsonKeys2int(x):
    if isinstance(x, dict):
        try:
            return {int(k):v for k,v in x.items()}
        except:
            pass #bad form
    return x

### Read In Merged DF and Decompression Dictionary
----
In order to reduce file size all categorical cells were converted into numbers and additionally compressed into a GZip file.

The next three cells

1. Reads in the numerical Dataframe
2. Reads the JSON file which contains the key:value pairs to decode the DataFrame
3. Executes a function that uses the JSON file to decompress/decode the numerical DataFrame.

df=pd.read_csv('./merged_df/merged_df.csv.gz', low_memory=False)

df.to_csv("./merged_df/merged_df.csv", index=False)

with open("./json_files/decompression_dictionary.json", "r") as json_file:
    decompression_dictionary = json.load(json_file, object_hook=jsonKeys2int)

df = compress_dataframe(df, decompression_dictionary)

### Quick Clean

df = reduce_large_strings(df)

#cast numerical features back to correct values
numerical_features = ['as_of_year', 'agency_code', 'loan_type', 'loan_purpose',
                      'owner_occupancy', 'loan_amount_000s', 'preapproval',
                      'action_taken', 'msamd', 'census_tract_number', 'applicant_ethnicity',
                      'co_applicant_ethnicity', 'applicant_race_1', 'co_applicant_race_1',
                      'applicant_income_000s', 'purchaser_type', 'denial_reason_1',
                      'denial_reason_2', 'denial_reason_3', 'hoepa_status', 'lien_status',
                      'population', 'minority_population', 'hud_median_family_income',
                      'tract_to_msamd_income', 'number_of_owner_occupied_units',
                      'number_of_1_to_4_family_units', 'latino', 'approve_bin']
dtypes = [int, int, int, int, int, float, int, int, float, float, int, int, int,int,
          float, int, float, float, float, int, int, float, float, float, float, float,
          float, int, int]
for col, dtype in zip(numerical_features, dtypes):
    df[col] = df[col].astype(dtype)

df.head()

**Baseline Score**
The baseline for the modeling of this study is the mean approval rate of the set.

df.isnull().sum().sort_values(ascending=False).head(15)

df.shape

df.dropna(axis=0,subset=['number_of_1_to_4_family_units','number_of_owner_occupied_units',\
                         'tract_to_msamd_income','minority_population','population',\
                         'hud_median_family_income','census_tract_number'], inplace=True)

Dropping the NAN entries that are not the denial reasons, as they make up only a small fraction of the dataset

df.shape

df['loan_purpose_name'].value_counts(ascending=False)

df=df[df['loan_purpose']==1].copy(deep=True)

Filtering by loan type, we wish to only look at the loan's that were categorized as 'Home purchase'.

print('Baseline mean approval of', round((df['approve_bin'].mean())*100,2),'%')

df[(df['state_name']=='New York') & (df['as_of_year']==2017)].approve_bin.mean()

## EDA Descriptive/Numerical Metrics

df.columns

List of features we have narrowed down, with the addition of the engineered features of latino and approved_bin

numerical_features = df.select_dtypes(include='number').columns.tolist()
print(f'There are {len(numerical_features)} numerical features:', '\n')
print(numerical_features)

numerical_features = df.select_dtypes(exclude='number').columns.tolist()
print(f'There are {len(numerical_features)} numerical features:', '\n')
print(numerical_features)

Due to the unfamiliarity of the dataset, it was chosen to use the data with both name tags and numerical encoding for the categories

df['as_of_year'].value_counts(normalize=True,ascending=False)

Based on our sorting patterns to reduce outliers/edge cases, the data based on year recorded normalized

df['state_name'].value_counts(normalize=True,ascending=False)

Florida and Texas make up ~65% of the dataset

df['msamd_name'].value_counts(normalize=True,ascending=False)

A quick look at the percentage of samples based on location.

df['applicant_sex_name'].value_counts(normalize=True)

Majority of applicants are male.

df['applicant_race_name_1'].value_counts(normalize=True)

Majority applicant race is white with 65% of the samples falling in that category and additionally 'Information not Provided' is the third highest category by sample number.

df[df['state_name']=='California'].groupby('applicant_race_name_1').approve_bin.mean()

df[df['state_name']=='Texas'].groupby('applicant_race_name_1').approve_bin.mean()

df[df['state_name']=='Florida'].groupby('applicant_race_name_1').approve_bin.mean()

df[df['state_name']=='New York'].groupby('applicant_race_name_1').approve_bin.mean()

Groupby splits for the states in question, we do see some trends with this EDA.  In all the sets Black/African American recorded applicant samples have lower approval numbers then the majority class (White applicants).  There is a significant number of applicants that do not have information provided for race and because this category is lower then the majority class for all states, it is very possible that this most likely contains more minority applicants then majority class applicants.  It should be noted that applicants that are recorded as Asian for California are at parity with White.

df.groupby('purchaser_type_name').approve_bin.mean()

This feature will not be useful in modeling, but it could be interesting to look at the demographic splits of the loan purchasers.

df.groupby('lien_status_name').approve_bin.mean()

df.groupby('msamd_name')[['approve_bin','minority_population','applicant_income_000s','hud_median_family_income']].mean().sort_values(by='approve_bin',ascending=False)

Florida lags behind specifically Miami Beach in loan approval rate and additionally by mean applicant income.  Specifically looking at Miami Beach, the lowest approval rate of the metro area's specifically we are looking at and additionally the highest mean percentage of minority population.  It should be noted that the next two highest minority population mean percentage metro areas are both California, so clearly applicant income is strong deciding factor as well, but one must consider that the standard of living in these locations is much different.  Adjustments would need to be made to accurate reflect applicant income effect more precisely.  It is interesting to see how Housing Urban Development median family income numbers stack up to the approval rates (the median family income per the data dictionary is per MS/AMD).  Applicant overall income does look to be a stronger correlation.

df[['loan_type_name','owner_occupancy_name']].value_counts()

Plotting on the loan type will be done, most of the loans are being taken out for owner's own housing.

df['co_applicant_ethnicity_name'].value_counts()

Most of the applicants do not have a co-applicant and even more so do not fall in the latino designation

df['co_applicant_race_name_1'].value_counts()

Again most applicant's are not filed with a co-applicant.  The vast majority of the samples are recorded as White.

df['hoepa_status_name'].value_counts()

Very few loans fall in the category of a loan that fall under the Home Ownership and Equity Protection Act.  The 200 entries are all approved so if further study was done with a larger number of samples that fell in this category in might be worth investigating the applicant parameters for a loan with this designation.

## Plots

plt.figure(figsize=(10, 7), facecolor="w")

sns.histplot(data=df, x="applicant_race_name_1", hue="approve_bin")
plt.xticks(rotation=45)

plt.title("Count of Applicants by Race:\nApproved, Not Approved", size=24)
plt.xlabel("Applicant Race", size=14)
plt.ylabel("Count", size=14);

plt.figure(figsize=(10, 7), facecolor="w")


sns.histplot(data=df[(df["approve_bin"] == 0)], x="applicant_race_name_1", hue="denial_reason_name_1",multiple='stack')
plt.xticks(rotation=45)

plt.title("Count of Applicants Denied by Race:\nSplit by Denied Reason", size=24)
plt.xlabel("Applicant Race", size=14)
plt.ylabel("Count", size=14);

plt.figure(figsize=(10, 10), facecolor="w")

ax = sns.countplot(data=df, x="loan_type_name", hue="approve_bin")

totals = df["loan_type_name"].value_counts().values
np.append(totals, totals)

for i, patch in enumerate(ax.patches):
    if i > 3:
        i = i-4
    percent = f"{(patch.get_height()/totals[i]):.2%}"
    x = patch.get_x() + patch.get_width()/6
    y = patch.get_height() / 2
    if i == 3:
        y = patch.get_height() + 1000
    ax.text(x, y, percent, color="k")
    
plt.title("Loan Approval by Type", size=24)
plt.xlabel("Loan Type", size=14)
plt.ylabel("Count", size=14);

fig = plt.figure(figsize=(10, 10), facecolor="w")

ax = sns.countplot(data=df, x="applicant_race_name_1", hue="approve_bin");
plt.xticks(rotation=45)

totals = df["applicant_race_name_1"].value_counts().values
np.append(totals, totals)

for i, patch in enumerate(ax.patches):
    if i > 6:
        i = i-7
    percent = f"{(patch.get_height()/totals[i]):.2%}"
    x = patch.get_x() + patch.get_width()/3
    y = patch.get_height() + 2000
    ax.text(x, y, percent, color="k", rotation=90)
    
plt.title("Loan Approval by Type", size=24)
plt.xlabel("Loan Type", size=14)
plt.ylabel("Count", size=14);

plt.figure(figsize=(10, 10), facecolor="w")

sns.countplot(data=df, x="loan_type_name", hue="action_taken_name");
plt.xticks(rotation=45)
plt.title("Loans by Type: Split by Result", size=24)
plt.xlabel("Loan Type", size=14);

| Race Name | Race Number  |
| --------- | ------------ |
| American Indian or Alaska Native | 1 |
| Asian | 2 |
| Black or African American | 3 |
| Native Hawaiian or Other Pacific Islander | 4 |
| White | 5 |
| Information not provided by applicant | 6 |
| N/A | 7 |
| No co-applicant | 8 |

fig, axes = plt.subplots(2,3, figsize=(18,10), facecolor="w")

fig.suptitle('Plots Set 1')

sns.histplot(ax=axes[0,0], data=df[df['loan_amount_000s']],x='loan_amount_000s')
sns.histplot(ax=axes[0,1], data=df,x='applicant_race_1')
sns.histplot(ax=axes[0,2], data=df,x='denial_reason_1')
sns.kdeplot(ax=axes[1,0], data=df,x='applicant_income_000s')
sns.kdeplot(ax=axes[1,1], data=df,x='minority_population')
sns.histplot(ax=axes[1,2], data=df,x='denial_reason_2')

plt.figure(figsize=(15,8), facecolor="w")
sns.voilinplot(data=df,x='applicant_race_1',y='applicant_income_000s',
               hue='approve_bin')
plt.title('Boxplots of tract percentage income', size=20);

# applicant race : 1 = American Indian or Alaska Native, 2 = Asian, 3 = Black or African American, 4 = Native Hawaiian or Other Pacific Islander
# applicant race : 5 = White, 6 = Information not provided by applicant, 7 = N/A, 8 = No co-applicant

#denial reasons : 1 = debt-income ratio, 2 = employment history, 3 = credit history, 4 = collateral, 5 = insufficient cash. 6 = Unverifiable information
#dneial reasons : 7 = unverifiable information, 8 = mortgage insurance denied, 9 = other

sns.set(style='whitegrid', palette="deep", font_scale=1.1, rc={"figure.figsize": [20, 10]})
sns.distplot(df['applicant_income_000s'], norm_hist=False, kde=False, bins=50, hist_kws={"alpha": 1}
).set(xlabel='Applicant Income Levels (Thousands)', ylabel='Count');
plt.xticks([x for x in range(int(df['applicant_income_000s'].min()),
                             int(df['applicant_income_000s'].max()),
                             50)])
plt.title('Application Income Levels');

## Plotly!

import plotly.graph_objects as go
import plotly.express as px
from plotly.subplots import make_subplots

fig=px.histogram(df[~(df['applicant_race_name_1']=='White') &\
                   (df['loan_amount_000s']<600)],
                 x='loan_amount_000s',
                 color='approve_bin',
                 nbins=200,
                 color_discrete_sequence=["red", "blue"],barmode='stack',
                title='Non-White Applicant Loan Amount Coded by Approved/Disapproved Count')
fig.update_layout(
#     xaxis = dict(
#     tickmode = 'linear',
#     tick0=0,
#     dtick=1
#     ),
    width=1400,
    height=450)
fig.show()

fig=px.histogram(df[(df['applicant_race_name_1']=='White') &\
                   (df['loan_amount_000s']<500)],\
                 x='loan_amount_000s',
                 color='approve_bin',
                 nbins=100,
                 color_discrete_sequence=["red", "blue"],
                 title='White Applicant Loan Amount Coded by Approved/Disapproved Count')
fig.show()

fig=px.histogram(df[~(df['applicant_race_name_1']=='White') &\
                   (df['loan_amount_000s']<500)],\
                 x='minority_population',
                 color='approve_bin',
                 nbins=100,
                 color_discrete_sequence=["red", "blue"],
                 title='Non-White Applicant Minority Population Tract % Coded by Approved/Disapproved Count'
                 )
fig.update_layout(
#     xaxis = dict(
#     tickmode = 'linear',
#     tick0=0,
#     dtick=1
#     ),
    width=1400,
    height=450)
fig.show()

fig=px.histogram(df[(df['applicant_race_name_1']=='White') &\
                   (df['loan_amount_000s']<500)],\
                 x='minority_population',
                 color='approve_bin',
                 nbins=100,
                 color_discrete_sequence=["red", "blue"],
                 title='White Applicant Minority Population Tract % Coded by Approved/Disapproved Count'
                 )
fig.show()

fig=px.histogram(df[~(df['applicant_race_name_1']=='White') &\
                   (df['applicant_income_000s']<300)],\
                 x='applicant_income_000s',
                 color='approve_bin',
                 nbins=100,
                 color_discrete_sequence=["red", "blue"]
                 )
xaxis = dict(
tickmode = 'linear',
tick0=40,
dtick=40)
fig.show()

fig=px.histogram(df[(df['applicant_race_name_1']=='White') &\
                   (df['applicant_income_000s']<300)],\
                 x='applicant_income_000s',
                 color='approve_bin',
                 nbins=100,
                 color_discrete_sequence=["red", "blue"],
                 )
fig.show()

df['latino'].value_counts(normalize=True)
