==> 04_all_states_model.ipynb <==
# All States
----
This contains the merged DataFrames of California, Texas, New York and Florida.

Data is limited to:
 - Years 2014-2017
 - People over the 20<sup>th</sup> percentile and under the 90<sup>th</sup> percentile by state
 - People in the few major Metro Areas by state

import json
import math
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.pipeline import Pipeline
from sklearn.decomposition import PCA
from sklearn.impute import SimpleImputer
from sklearn.compose import ColumnTransformer
from sklearn.neighbors import KNeighborsClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.preprocessing import OneHotEncoder, MinMaxScaler, StandardScaler
from sklearn.model_selection import train_test_split, GridSearchCV, cross_val_score, KFold

def compress_dataframe(df, dictionary):
    df = df.copy(deep=True)
    
    for col in dictionary.keys():
        df[col] = df[col].map(dictionary[col])
    return df

def reduce_large_strings(df):
    to_replace = ["Information not provided by applicant in mail, Internet, or telephone application",
                  "Native Hawaiian or Other Pacific Islander",
                  "One-to-four family dwelling (other than manufactured housing)"]
    replace_with = ["Information not Provided",
                    "Native Hawaiian/Pacific Islander",
                    "1-4 Family House"]
    for col in df.columns:
        if df[col].dtype == object:
            for i, _ in enumerate(to_replace):
                df[col] = np.where(df[col] == to_replace[i], replace_with[i], df[col])            
    return df

# Code from Murmel on Stack Exchange
# https://stackoverflow.com/questions/1450957/pythons-json-module-converts-int-dictionary-keys-to-strings
def jsonKeys2int(x):
    if isinstance(x, dict):
        try:
            return {int(k):v for k,v in x.items()}
        except:
            pass #bad form
    return x

### Read In Merged DF and Decompression Dictionary
----
In order to reduce file size all categorical cells were converted into numbers and additionally compressed into a GZip file.

The next three cells

1. Reads in the numerical Dataframe
2. Reads the JSON file which contains the key:value pairs to decode the DataFrame
3. Executes a function that uses the JSON file to decompress/decode the numerical DataFrame.

df=pd.read_csv('./merged_df/merged_df.csv.gz', low_memory=False)

df.to_csv("./merged_df/merged_df.csv", index=False)

with open("./json_files/decompression_dictionary.json", "r") as json_file:
    decompression_dictionary = json.load(json_file, object_hook=jsonKeys2int)

df = compress_dataframe(df, decompression_dictionary)

### Quick Clean

df = reduce_large_strings(df)

#cast numerical features back to correct values
numerical_features = ['as_of_year', 'agency_code', 'loan_type', 'loan_purpose',
                      'owner_occupancy', 'loan_amount_000s', 'preapproval',
                      'action_taken', 'msamd', 'census_tract_number', 'applicant_ethnicity',
                      'co_applicant_ethnicity', 'applicant_race_1', 'co_applicant_race_1',
                      'applicant_income_000s', 'purchaser_type', 'denial_reason_1',
                      'denial_reason_2', 'denial_reason_3', 'hoepa_status', 'lien_status',
                      'population', 'minority_population', 'hud_median_family_income',
                      'tract_to_msamd_income', 'number_of_owner_occupied_units',
                      'number_of_1_to_4_family_units', 'latino', 'approve_bin']
dtypes = [int, int, int, int, int, float, int, int, float, float, int, int, int,int,
          float, int, float, float, float, int, int, float, float, float, float, float,
          float, int, int]
for col, dtype in zip(numerical_features, dtypes):
    df[col] = df[col].astype(dtype)

df.isnull().sum().sort_values(ascending=False).head(15)

df.shape

df.dropna(axis=0,subset=['number_of_1_to_4_family_units','number_of_owner_occupied_units',\
                         'tract_to_msamd_income','minority_population','population',\
                         'hud_median_family_income','census_tract_number'], inplace=True)

Dropping the NAN entries that are not the denial reasons, as they make up only a small fraction of the dataset

df.shape

| Race Name | Race Number  |
| --------- | ------------ |
| American Indian or Alaska Native | 1 |
| Asian | 2 |
| Black or African American | 3 |
| Native Hawaiian or Other Pacific Islander | 4 |
| White | 5 |
| Information not provided by applicant | 6 |
| N/A | 7 |
| No co-applicant | 8 |

#### in this model we are looking at home purchases

df = df[df["loan_purpose_name"] == "Home purchase"].copy(deep=True)

#### the not applicable sex name for applicant had almost 100% loan rate so this was deemed to be not good for the model

df=df[~(df['applicant_sex_name']=='Not applicable')].copy(deep=True)

### initial test run for new york - uncomment if you want smaller sample (NY Metro)
### df = df[(df["as_of_year"] == 2017) & (df["state_name"] == "New York")]

df.shape

df.isna().sum()

#### a copy of the dataset just in case
df_copy = df.copy()

#### hard coded drops of columns we don't need for sure
df.drop(columns = ['respondent_id',
                   'denial_reason_name_1','denial_reason_name_2','denial_reason_name_3',
                   'denial_reason_1','denial_reason_2','denial_reason_3','action_taken_name','action_taken',
                  'purchaser_type_name','lien_status_name','county_name','agency_abbr',
                  'agency_code','msamd','owner_occupancy','loan_purpose_name','lien_status','purchaser_type','hoepa_status_name'],
        inplace = True)

#### grab the columns that have "name" in the ... name and then drop those if there is a numerical analog
name_analog = [x.replace("_name","") for x in df.columns if x.find("name") > 0]
name_analog_clean = [x for x in name_analog if x in df.columns]

df.drop(name_analog_clean, axis=1, inplace=True)

df.columns

#purchaser type -> leakage -> since it implies a loan was originated
#df.groupby('purchaser_type').approve_bin.mean()

Purchaser_type_name and lien_status_name are features that leak into our target variable so we needed to drop them from our model

print('Baseline mean approval of', round((df['approve_bin'].mean())*100,2),'%')

X_train, X_test, y_train, y_test = train_test_split(df.drop(columns = 'approve_bin', axis=1), 
                                                    df.approve_bin, test_size=0.75, random_state=42)

numerical_features = X_train.select_dtypes(include='number').columns.tolist()
print(f'There are {len(numerical_features)} numerical features:', '\n')
print(numerical_features)

categorical_features = X_train.select_dtypes(exclude='number').columns.tolist()
print(f'There are {len(categorical_features)} categorical features:', '\n')
print(categorical_features)

numeric_pipeline = Pipeline(steps=[
    ('impute', SimpleImputer(strategy='mean')),
    ('scale', StandardScaler())
])

categorical_pipeline = Pipeline(steps=[
    ('impute', SimpleImputer(strategy='most_frequent')),
    ('one-hot', OneHotEncoder(handle_unknown='ignore', sparse=False))
])

full_processor = ColumnTransformer(transformers=[
    ('number', numeric_pipeline, numerical_features),
    ('category', categorical_pipeline, categorical_features)
])


#### Logistic : First Pass

logreg = LogisticRegression(random_state=42)

logreg_pipeline = Pipeline(steps=[
    ('preprocess', full_processor),
    ('clf', logreg)
])

logreg_grid = [{'clf__penalty': ['l2'],
                'clf__C': [1.0,0.5,0.1],
                'clf__solver': ['sag'],
                'clf__max_iter': [1000]
               }] 

logreg_cv = GridSearchCV(logreg_pipeline, logreg_grid, 
                      cv=10, 
                      scoring='accuracy',
                        n_jobs=-1)

logreg_cv.fit(X_train,y_train)

print("tuned hpyerparameters :(best parameters) ",logreg_cv.best_params_)
print("accuracy :",logreg_cv.best_score_)

coeff_val = logreg_cv.best_estimator_.named_steps.clf.coef_


best_cat = logreg_cv.best_estimator_.named_steps['preprocess'].transformers_[1][1].named_steps['one-hot'].get_feature_names(categorical_features)

coeff = numerical_features

coeff.extend(best_cat)

coeff_zip = zip(coeff,coeff_val[0])

c_l = list(coeff_zip)

df_fin =  pd.DataFrame(c_l,columns = ['Coefficient', 'Value'])

df_fin['abs_Val'] = df_fin['Value'].map(lambda x : abs(x))

df_fin.sort_values(by = 'abs_Val', ascending  = False)[:50]

## PCA

#### reseting the numerical and categorical features and pipelines because the PCA pipline below did not originally like the 
#### coefficient exploration above 
numerical_features = X_train.select_dtypes(include='number').columns.tolist()
print(f'There are {len(numerical_features)} numerical features:', '\n')
print(numerical_features)

categorical_features = X_train.select_dtypes(exclude='number').columns.tolist()
print(f'There are {len(categorical_features)} categorical features:', '\n')
print(categorical_features)

numeric_pipeline = Pipeline(steps=[
    ('impute', SimpleImputer(strategy='mean')),
    ('scale', StandardScaler())
])

categorical_pipeline = Pipeline(steps=[
    ('impute', SimpleImputer(strategy='most_frequent')),
    ('one-hot', OneHotEncoder(handle_unknown='ignore', sparse=False))
])

full_processor = ColumnTransformer(transformers=[
    ('number', numeric_pipeline, numerical_features),
    ('category', categorical_pipeline, categorical_features)
])

pca = PCA()
log_reg = LogisticRegression()

log_reg_pipe = Pipeline(steps=[
    ('preprocess', full_processor),
    ('pca',pca),
    ('model', log_reg)
])

param_grid = {
    "pca__n_components": [5, 10, 15, 20],
    "model__C": np.logspace(-4, 4, 4)
}
search = GridSearchCV(log_reg_pipe, param_grid, n_jobs=2)
search.fit(X_train, y_train)
print("Best score (CV score=%0.3f):" % search.best_score_)
print("Best params:", search.best_params_)


### PCA quick analysis

# ~80 % explained by the first 10 components!
pca.fit(full_processor.fit_transform(X_train_2))
sum(pca.explained_variance_ratio_[:15])

pca1_evec = pca.components_[0]
for weight, event in zip(pca1_evec, df.iloc[:,1:].columns):
    print(event, weight)

X = df.drop(columns = 'approve_bin', axis=1)
pca_df = pd.DataFrame(pca.transform(full_processor.fit_transform(X)),
                      columns=['pr_comp_' + str(i) for i in range(full_processor.fit_transform(X).shape[1])])
pca_df.head()

pca_df['approve_bin'] = list(df.approve_bin)

lr = LogisticRegression(random_state=42)

scores = cross_val_score(lr,
                         pca_df.iloc[:, :15],
                         pca_df.iloc[:, -1],
                         cv=KFold(n_splits=5,
                                  shuffle=True,
                                  random_state=42)   
                        )
round(scores.mean(), 2), '+-', round(2 * scores.std(), 2)

knn = KNeighborsClassifier()
scores = cross_val_score(knn,
                         pca_df.iloc[:, :15],
                         pca_df.iloc[:, -1],
                         cv=KFold(n_splits=5,
                                  shuffle=True,
                                  random_state=73))
round(scores.mean(), 2), '+-', round(2 * scores.std(), 2)

## Random Forrest 

pca = PCA()
rf = RandomForestClassifier()

rf_pipe = Pipeline(steps=[
    ('preprocess', full_processor),
    ('pca',pca),
    ('model', rf)
])

param_grid = [{"pca__n_components": [15],
               "model__n_estimators": [100]
               }] 

search = GridSearchCV(rf_pipe, param_grid, cv=5, scoring='accuracy', n_jobs=-1)

search.fit(X_train,y_train)

print("tuned hpyerparameters :(best parameters) ",search.best_params_)
print("accuracy :",search.best_score_)

#https://towardsdatascience.com/grid-search-for-model-tuning-3319b259367e
#https://towardsdatascience.com/polynomial-regression-with-a-machine-learning-pipeline-7e27d2dedc87
#https://scikit-learn.org/stable/auto_examples/compose/plot_digits_pipe.html
#https://medium.com/@andymdc31/using-pca-in-a-machine-learning-pipeline-b6fe3492b1b9
